+++ changelog.md
@@ -3,6 +3,7 @@
 - 18:20 - Front + Back: "Download" button on the frontend is more or less working.
 - 18:30 - Backend: `Download.bookmark` type changed to `OneToOneField`.
 - 18:40 - Front + Back: refactored according to new `Download.bookmark` type.
+- 19:40 - Backend: added retries to `download_from_youtube()` because of `pytube` bug.
 
 ## 22.04.2023, Sat